feat(react-ui-ag): Modify MobileListing and stories, theme, and test for it
affects: @rentpath/react-ui-ag
