+++ packages/react-ui-ag/src/Listings/index.js
@@ -1,8 +1,7 @@
 export { default as MobileMapListing } from './MobileMapListing'
 export { default as SingleFamilyMobileMapListing } from './SingleFamilyMobileMapListing'
+export { default as MobileListing } from './MobileListing'
 export { default as DesktopMapPinListing } from './DesktopMapPinListing'
 export { default as Listing } from './Listing'
 export { default as DesktopListing } from './DesktopListing'
 export { default as SingleFamilyDesktopListing } from './SingleFamilyDesktopListing'
-export { default as MobileListing } from './MobileListing'
-export { default as SingleFamilyMobileListing } from './SingleFamilyMobileListing'
