+++ stories/ag/Listings/MobileListing.js
@@ -3,42 +3,11 @@ import { action } from '@storybook/addon-actions'
 import { MobileListing } from 'react-ui-ag/src'
 import { boolean } from '@storybook/addon-knobs'
 import ratingTheme from '../../theme/core/RatingBar.css'
-
-const baseListing = {
-  banners: ['$ Coupon'],
-  bedrooms: '1-3 Beds',
-  bathrooms: '2 Bathrooms',
-  name: 'Awesome Property!',
-  city: 'Great Town',
-  state: 'YA',
-  price: '$1170+',
-  address: '3921 Tugaloo River Drive Northside Park Washington Parkway',
-  unitLevelAvailability: '• 8 units left',
-  availability: 'Available Now',
-  rating: {
-    score: 4,
-    label: '20',
-  },
-  photos: [
-    {
-      path: 'imgr/2576db62ffa153ebef00317a5c68a368/',
-      caption: 'test 1',
-    },
-    {
-      path: 'imgr/d56984e959a3feb1235f85ee202a0fc6/',
-      caption: null,
-    },
-    {
-      path: 'imgr/fd972eb03a0463c484580349ad5177b7/',
-      caption: null,
-    },
-  ],
-  url: '/apartments/Georgia/Lawrenceville/Sugarloaf-Summit/100029616/',
-  phone: '404-378-1428',
-}
+import baseListing from './baseListing'
 
 const props = {
   listing: baseListing,
+  sponsoredText: 'Sponsored',
   onClick: () => action('click')('listing cell click'),
   favoriteButton: {
     onClick: (listing, value) => action('click')('favorite toggled to ', value),
@@ -46,7 +15,7 @@ const props = {
   },
   photos: {
     server: 'https://image.rent.com/',
-    dimensions: '280-120',
+    dimensions: '351-160',
   },
   ratings: {
     className: ratingTheme.themedRating_starAgMobile,
@@ -54,15 +23,12 @@ const props = {
   },
   ctaButtons: [
     {
-      valueLocation: 'phone',
-      onClick: () => action('click')('Phone CTA action'),
-      className: 'phone',
-      itemProp: 'telephone',
+      children: 'Check Availability',
+      onClick: () => action('click')('contact CTA action'),
     },
     {
-      children: 'send an email',
-      onClick: () => action('click')('contact CTA action'),
-      className: 'contact',
+      children: 'Contact Property',
+      onClick: () => action('click')('coupon CTA action'),
     },
   ],
 }
@@ -71,13 +37,15 @@ const singleFamilyProps = {
   ...props,
   listing: {
     ...props.listing,
+    banners: [],
     singleFamily: true,
   },
-  ctaButtons: [{
-    onClick: () => action('click')('More Information action'),
-    className: 'phone',
-    children: 'More Information',
-  }],
+  ctaButtons: [
+    {
+      children: 'Check Availability',
+      onClick: () => action('click')('contact CTA action'),
+    },
+  ],
 }
 
 export const ExampleMobileListing = () => {
@@ -85,6 +53,7 @@ export const ExampleMobileListing = () => {
   const listing = {
     ...baseListing,
     isFavorited: boolean('listing.isFavorited', false),
+    sponsored: true,
   }
   return (
     <MobileListing
@@ -95,7 +64,7 @@ export const ExampleMobileListing = () => {
   )
 }
 
-export const ExampleSingleFamilyMobileListing = () => {
+export const ExampleMobileListingSingleFamily = () => {
   const isActive = boolean('isActive', true)
   const listing = {
     ...singleFamilyProps.listing,
