+++ stories/ag/Listings/baseListing.js
@@ -1,0 +1,43 @@
+const baseListing = {
+  banners: ['Coupon'],
+  photos: [
+    {
+      path: 'imgr/2576db62ffa153ebef00317a5c68a368/',
+      caption: 'test 1',
+    },
+    {
+      path: 'imgr/d56984e959a3feb1235f85ee202a0fc6/',
+      caption: null,
+    },
+    {
+      path: 'imgr/fd972eb03a0463c484580349ad5177b7/',
+      caption: null,
+    },
+  ],
+  price: '$1170+',
+  bedrooms: '1-3 Beds',
+  bathrooms: '• 1-2 Baths',
+  unitLevelAvailability: '12 Units Available',
+  lastUpdated: '3hrs ago',
+  availability: 'Available Now',
+  name: 'Really Really Long Name Property',
+  address: '3921 Tugaloo River Drive Northside Park Washington Parkway',
+  rating: {
+    score: 4,
+    label: '(20)',
+    count: 62,
+    reviews: 9,
+  },
+  phone: '(404) 770-5555',
+  location: {
+    addressLine1: '3921 Tugaloo River Drive Northside Park Washington Parkway',
+    city: 'Beckwood',
+    latitude: 33.8362,
+    longitude: -84.3426,
+    state: 'Georgia',
+    zip: '30319',
+  },
+  url: '/apartments/Georgia/Lawrenceville/Sugarloaf-Summit/100029616/',
+}
+
+export default baseListing
