+++ stories/ag/Listings/index.js
@@ -1,4 +1,4 @@
 export { ExampleMobileMapListing, ExampleSingleFamily } from './MobileMapListing'
 export { ExampleDesktopMapPinListing } from './DesktopMapPinListing'
 export { ExampleDesktopListing, ExampleSingleFamilyDesktopListing } from './DesktopListing'
-export { ExampleMobileListing, ExampleSingleFamilyMobileListing } from './MobileListing'
+export { ExampleMobileListing, ExampleMobileListingSingleFamily } from './MobileListing'
